requirements.txt
aiohttp==3.9.3 
uvicorn==0.34.2 
jinja2==3.1.6
beautifulsoup4==4.12.3
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
pytest==8.3.5
fastapi==0.115.12 
